Make BACKGROUND_TIMER_SETTING disable creating worker
