+++ cofi-share/src/main/java/com/omelan/cofi/share/DataStoreKeys.kt
@@ -7,6 +7,7 @@ import androidx.datastore.preferences.core.stringPreferencesKey
 val STEP_SOUND_ENABLED = booleanPreferencesKey("ding_enabled")
 val STEP_VIBRATION_ENABLED = booleanPreferencesKey("step_vibration_enabled")
 val COMBINE_WEIGHT = stringPreferencesKey("combine_weight")
+val BACKGROUND_TIMER_ENABLED = booleanPreferencesKey("background_timer_enabled")
 
 const val STEP_SOUND_DEFAULT_VALUE = true
 const val STEP_VIBRATION_DEFAULT_VALUE = true
